extract album version into separate file
